File > Export > Image to file... now available
Thank you to @hi5 for discussing this feature in #513.

Images can now be "exported" to standalone image files (in any supported format).  How does this differ from a normal "Save" or "Save as" event?

The main difference is that an export event does not modify the current image's save state, filename, or target format.  So if you load "image.jpg", make changes to it, then *export* it to "new image.png", the image still appears as "image.jpg" in PhotoDemon.  It also will prompt you about unsaved changes if you try to close the image (because the export is not considered a "save" event).  If you click "Save", it will overwrite the original "image.jpg", not "new image.png".

This is primarily useful when working with a multi-layer image that is ultimately exported to a single-frame format.  You can save/load/work with the image in PhotoDemon's native PDI format, and export it periodically to a flat format without worrying about to later forget to save it (losslessly) back to PDI.

The new Export menu also explicitly splits out exporting still vs animated formats, so you could theoretically also use this menu to save an animated image to a flat file (why would you want to do this? idk).

Next up in this branch: a new File > Export > Layers to Files.  This will (hopefully) mirror a similar function in Photoshop, and allow the user to export all layers in an image to individual files.