New framework for automatic plugin updates
Relates to #534 .  Thank you to @mizka76 for reporting this problem.

PhotoDemon can work with AVIF files, but it doesn't ship the required 3rd-party plugin (libavif) because libavif is *huge* - bigger than PhotoDemon itself!  Instead, PD offers to download libavif on-the-fly when a user first interacts with an AVIF file.

This works well, but there is no additional mechanism to update libavif if the installed version falls out of date with the latest version PD supports.  This commit fixes the problem with a new plugin update feature.

PD now checks plugin versions when working with AVIF files, and if it sees that the installed version is less than the latest supported version, PD will now offer to download and configure a new copy of libavif.  This of course requires an internet connection, and the user can say NO to keep working with their existing plugin copy.  But if they say YES, PD will download the latest libavif, decompress and install it locally (to PD), and continue with the action that initially triggered the update warning (usually loading or saving an AVIF file).

As part of this work, I also rewrote libavif's version-checking and download code to use a new generalized set of update functions.  I should be able to migrate other on-demand image formats (like JPEG XL) to this same codebase, which would extend ongoing automatic updates to those formats as well.

I also updated the automatically downloaded libavif version to the latest 1.0.4 while here (previously PD used 1.0.1).

Still TODO is updating localizations for the new update feature - this will happen shortly!