PDF import now available
A lot of work is still TODO (including import UI and all localization work), but with this commit it is possible to import a PDF into PhotoDemon.  Each page is successfully loaded into its own layer.  PDF resolution is hard-coded for now; this will be user-controllable once a UI is built.

As usual, I plan to get PDF import features 100% completed before tackling PDF export.