Remove "Overwrite" from blend mode UIs
A proper "Overwrite" blend mode has challenging complications with paint brushes (because we need to pass the current brush *mask* to the compositor, so it knows what region to overwrite) so I don't want to tackle it right this second.  I'd rather address some other much-needed fixes and improvements first - then return to Overwrite once I've not got so many pressing things to solve.