(try to) bring PD to the foreground when a new image is loaded...
...from an external window (like Windows Explorer).

This is only relevant in single-session mode.  If PD is in the background, and the user loads an image from Windows Explorer, PD will now try to bring PD to the foreground.  This is *not* guaranteed to happen (due to system policies), but if it fails, Windows typically compensates by flashing the app icon in the taskbar.