Finish UI for toggling "snap to canvas edge" behavior
A menu is now provided (`View > Snap to canvas edge`), and custom snap distance is changeable via `Tools > Options > Interface`.

A lot of code is required to provide a UI for this feature, since snap settings need to be persistent between sessions, but still easily toggleable in a top-level menu.

Next up: actually implementing the snap feature (sigh).