Snap to Centerlines now works!
For layers *and* selections!  This makes it trivial to snap a smaller layer to the center of the canvas (on either an edge of the smaller layer, or the center of the smaller layer).

In a rare bit of good engineering, I can add arbitrary snap targets whenever I want (e.g. next up: layer boundaries), and all objects automagically pick up the "snap" ability for those new targets.  No snap engine modifications are required when new snap targets get added.