Much smarter interaction between "snap" and "lock aspect ratio" settings
Locking a layer's aspect ratio *and* snapping it to nearby objects produces complex interactions (because enforcing one often breaks the other).

This commit adds a clear hierarchy to the process.  Locked-aspect-ratio objects are first snapped to position, then aspect ratio is re-calculated and the lock is performed (generating new width/height, potentially), then subsequent snaps on the other axis are enforced IFF aspect ratio can be maintained (in most cases they cannot).

This produces much better behavior when click-drag resizing a layer with locked aspect ratio using the move/size tool, on an image with many layers, while snapping to layers is enabled.