Status bar: add % as an available unit of measurement
This allows things like mouse coordinates, ruler notches, distances, layer dimensions, and selection sizes to be displayed as a percentage of total image size, which should be very helpful for alignment purposes.

(A lot of random tools and UI bits were affected by this change, but all measurement conversions should now be using a central converter that standardizes things like significant digits on a per-unit basis.  This greatly improves consistency and cuts a bunch of redundant unit display code.)