HEIF/HEIC export: multi-frame exports now supported
The HEIF/HEIC export dialog now allows you to export just the composited image, or one-frame-per-layer.  This allows for successful round-tripping of multi-frame HEIC images.

I've also switched PD's default extension for exporting these images to HEIC instead of HEIF since the HEVC codec (h.265) is used, and if I'm reading the spec correctly, HEIC indicates HEVC encoding.

With this commit, I consider HEIF/HEIC support roughly complete for now.  Note however that vcpkg's version of libheif trails the official version, and as noted in PD's code some known libheif bugs that currently require workarounds may go away as new libheif builds become available - so I'll keep an eye on vcpkg in the coming weeks to see if it picks up v1.18.x (instead of 1.17.6, which is the latest available as of this commit).