Hotkey edits now save across sessions
You can now...

- Edit hotkeys (`Tools > Keyboard shortcuts...`)
- Hit OK (in the `Keyboard shortcuts` dialog)
- Use new hotkeys
- Close PhotoDemon
- Start PD again. your hotkey edits are all present and active!

This means hotkey edits are now synced between the editor, hotkey manager, and menu manager.  This was a *lot* more work than I expected.

There's still quite a bit of clean-up to do, and some small issues to resolve (like text and dialogs when attempting to save a hotkey edit that's duplicated elsewhere), but the meat of this feature is all in place.