PCX import: new internal engine
This should fix crashes and incompatibility with some valid PCX files currently loaded via FreeImage.  (Photoshop includes PCX as a native filetype, so we want to handle it correctly too.)

As of this commit, 24-bpp and 8-bpp PCX files are correctly loaded into PD using new, all-in-house code.

The PCX spec available here was invaluable for this project:
https://www.fileformat.info/format/pcx/spec/a10e75307b3a4cc49c3bbe6db4c41fa2/view.htm

1, 2, and 4 bpp PCX support will follow soon.