PCX import: attempt recovery of broken files
A little bit of code lets us rescue some "broken" PCX files found in the wild.  It's not foolproof (for safety reasons - RLE runs that could overflow more than one line are intentionally rejected) but it does help, and it means PD can recover some images that Photoshop cannot, which is kinda neat.

AFAIK, this is the end of my work on PCX import.  I feel good about the format's new coverage and since PCX is a native Adobe Photoshop file format, it's great that we can finally cover it correctly too.