On failed startup, create a small, "emergency" debug log
I was testing today on a fresh, zero-updates Win 7 VM and PD failed to start due to some 3rd-party library issues.  (Updated Win 7 installs work fine, so this is an extreme edge-case.)  I was curious about what specifically failed to load and some extra debug logging was helpful.

If users have issues in the future with PD startup issues, this should also help me nail down if the problem is on my end or theirs.