Crop tool: new "allow enlarging" toggle
The code for this feels much uglier than it should.  I want to think of a cleaner way to accomplish it, but nothing's occurring to me right now.  I'll take another look tomorrow!

Regardless of ugliness, the crop tool now enforces image boundaries *unless* the "allow enlarging" toggle is set.