Fixes for crashes with images around ~2 GB in size
PD was using ints to estimate allocation sizes when writing data to disk; as file sizes grow this is largely unnecessary (it's just a performance hint to try and get close to actual size up-front).

If an image expects to require more than 2 GB in storage space, we now skip estimating and just let it fill space as it requires.  This prevents crashes when the estimation of file size approaches or exceeds 2 GB.