Adjustments > Histogram: unify histogram rendering with other dialogs...
...that use histogram overlays (Curves, Levels).

In the dedicated "Display Histogram" dialog I've also added a toggle for using the moving average; this is used in Curves and Levels dialogs to display a "cleaner" overlay (since those dialogs are really about general distribution, not specific spikes).  In the dedicated histogram window, however, the user can now turn that behavior on/off because if they're using that tool, they're likely looking for something specific.

Thank you to Heiner D for catching and reporting this, and also for his many other suggestions and bug reports in recent days!