Implement "Replace" (Overwrite) blend mode
Like other photo editors, this blend mode will only be available on certain tool dialogs.

It primarily exists to enhance adjustments and effects, by providing a way to either replace the contents of the current image with your adjustments (default), or to "layer" those adjustments onto the image using variable blend modes and/or opacity.  This shortcuts the process of duplicating a layer, applying edits to that duplicate, then merging the results down with a custom layer blend mode + opacity.

Because blend mode and opacity UI toggles need to be manually added to relevant adjustments and effect tools, availability of this feature will roll out slowly in the coming weeks.

(Today, I mostly wanted to get the basic code merged so I can start working on it intermittently while continuing work on the crop tool and other issues.)