Bad image file extensions: warn user, and offer to fix
Relates to #644 .  Thank you to @hi5 for this suggestion!

When PhotoDemon detects a mismatch between an image's file extension and its actual file contents, it will now offer to rename the file with a correct extension.

(As with other popup dialogs, this behavior is disabled during batch processes, when loading multiple images at once, and when loading an image that didn't originate from disk - like images pasted from the clipboard.)

The popup warning and offer to fix do require some new localizations.

Please let me know if you encounter any weirdness with this feature.  It *should* only trigger when a file extension explicitly matches a known filetype *other* than the current file's, but there are a lot of filetypes out there and it's possible some extension combinations aren't fully caught.