Images and layers can now be suspended to disk
PD already supports suspending images and layers to compressed memory streams.  It uses this feature a *lot* "behind-the-scenes" to cut object counts (user and GDI) and reduce memory usage, particularly in UI elements.

Suspending pixel data to memory is less helpful on user images, however, because photos don't losslessly compress very well so they can still consume a *lot* of memory when suspended (plus we need a temporary buffer to store the compression results, so this can sometimes use *more* memory if used unwisely).

To that end, I've finally added a way to transparently suspend an entire layer or image to disk.  This feature uses fast lz4 compression to dump pixel data to a temp file, and the data is left there until the image's pixel data is accessed again - at which point the pixel data is silently restored from file, and the temp file deleted.  The caller doesn't need to know anything about suspend state; it's all handled silently "behind-the-scenes", so you can suspend data willy-nilly and everything still works.

I want to use this to start implementing "cry for help" scenarios in memory-sensitive functions, where a function can "cry for help" if a memory allocation fails, and PD will attempt to aggressively suspend image data wherever it can to try and make that allocation happen.

Such a feature is still TBD, but it will rely on this new "suspend to disk" feature to work.