When loading an image, other open images may now be auto-suspended...
... to disk to alleviate memory pressure.

PhotoDemon has long supported the ability to "suspend" a pixel surface to a compressed memory stream.  This reduces a surface's memory requirements, and frees up other system-limited resources (like user and GDI object handles).  Pixel surfaces are automatically un-suspended if/when a caller interacts with them again.

As of this commit, PhotoDemon now goes a step further and checks memory usage when loading a new image.  If doing so would put PD over 80% of its system-allotted memory usage (or 80% of the ~2 GB mark typically allotted to an x86 app), it will now suspend other open images to disk before attempting to load the image.  This *should* decrease the likelihood of out-of-memory crashes when working with many images simultaneously.

This behavior requires no interaction on the end-user's part, and suspended images are automatically un-suspended if/when the user interacts with them.

I would like to implement this behavior elsewhere, but adding it requires a lot of testing to ensure everything works OK after, so I'm starting here and will extend the feature to more places in the coming days.