Fix "offer to fix bad file extension" when drag/drop is used
The problem was a drag/drop one all along 😫

PhotoDemon suspends popup messages when multiple images are loaded.  This is to prevent a scenario where the user drag/drops a huge list of files and they all have problems, and the UI gets stuck in an endless stream of popups.  (A better solution would be a "cancel" button on all popup dialogs that cancels the entire load operation for *all* images - I'll look into this.)

Anyway, drag/drop was *always* requesting the suppression of warning popups, even when just one file was being drag/dropped.  Sigh.

On the plus side, it's an easy fix!

Relates to #644.  Thank you to @hi5 for patient help with this one!