AVIF: report libavif failure messages directly to user
Relates to #651.  Thank you to @manfromarce for catching and reporting!

libavif operates in "strict" mode by default, rejecting non-conforming images (and nicely reporting why they fail).  In this commit, I've switched to explicitly requesting "non-strict" behavior, which allows AVIF files created with older libavif versions to still work.

If an AVIF file is still rejected by libavif, the error is now reported directly to the user, so they know it's a problem with the source image and *not* with PhotoDemon.