Automatic memory reduction now triggers when creating pixel surfaces
The most likely place for PhotoDemon to hit an out-of-memory error is when initializing a pixel surface.  (They are the largest objects in the project, by far.)

If PD now tries to initialize a pixel surface and fails, it will aggressively attempt to suspend other open pixel surfaces to disk.  After it suspends as many other surfaces as it can, PD will then attempt to initialize the problematic surface again.

This allows PD to effectively load or create an "unlimited" number of images, provided you have disk space to store all the suspended images.  (PhotoDemon does compress these surfaces if it has enough memory to do so, so with a large enough hard drive, we're talking *millions* of open images.)

This feature is awesome but it's not capable of miracles.  If a single very enormous image is the problem, there's nothing PD can offload to fix the problem and it will just crash (sorry).  That's life as a 32-bit app.

But in situations where multiple images are loaded, or an operation requires multiple temporary copies of a surface for processing an adjustment or effect, this will allow PhotoDemon to succeed where previously it would just crash.

Hopefully this resolves #640, but without follow-up from that user, I have no way of knowing...