Fix app icon under Alt+Tab Terminator
Fixes #486.  It took years, but I eventually found the source of this damn bug!!

Thank you to Alexander, the developer of Alt+Tab Terminator, for his email correspondence on this issue.  And thank you to @hi5 for catching and reporting this problem initially!

(Yes, the problem was a typo on my part.  The bug was obscured because PD typically uses the active image as the app icon, but early in startup, it also sets a default app icon to the hidden top-level "ThunderMain" window used by VB6.  It was *that* function that had a typo where the large and small app icons were being switched.  Windows always pulls the current main window icon, so I couldn't reproduce the bug there, but Alt+Tab terminator uses only the icon from the top-level window, hence the 16x16 icon instead of the 32x32 one.)