Tools > Options: finish (?) overhauling underlying panel engine
Each `Tools > Options` panel now exists as a separate window, and at run-time I simply "stitch" individual panels onto the parent Options window if/when the user interacts with them.  As noted in previous commits, this greatly decreases load-time of this dialog, greatly simplifies the underlying code (by organizing it per-panel), and perhaps most importantly, it makes it much easier for me to expand Options panels in the future.

(Note also that translation files are reordered a lot by this commit, but nothing in them has actually changed!)