Support for custom user fonts: proof of concept
Relates to #499

This commit adds a user font folder (/Data/Fonts) to PD's user folder, and allows the user to invisibly use fonts in this folder as if they were installed system fonts.  The fonts are *not* installed and are only available to PD.

This commit was made complicated by the way PD uses a mix of font engines to achieve its various typography effects.  Workarounds for both GDI (simple!) and GDI+ (not simple!) and their interop (kill me now!) were required.  I *think?* all combinations are working correctly?

Next up is a UI for the user to add their own font folder locations.  This will be part of the Tools > Options > Fonts panel.