Tools > Options: dedicated metadata panel
Organizing interfaces is hard.  For days I've gone back and forth on how to organize the Tools > Options dialog, and this is the last change I intend to make for awhile.  Apologies for the ongoing language tweaks, translators!

I mostly hope the new organization makes it easier for users to find settings they need, and easier for me to maintain those settings going forward.  There are a lot of new categories in the current layout, but as a trade-off, individual category panels are significantly less dense.  I hope this is the right decision!