Menu mnemonics: check both localized *and* en-US text with IsAlphaCharW
PD's menu mnemonic algorithm ignores menus that don't contain alpha-numeric characters (like menu separators).  Unfortunately, this relies on a Windows API whose behavior changes based on the active language:

https://learn.microsoft.com/en-us/windows/win32/api/winuser/nf-winuser-ischaralphaw

For a portable app like PD, it's preferable to make decisions using the app language instead of the system language.  With this commit, PD will do a second, fallback detection on IsAlphaCharW on the en-US text to be more certain if a menu item warrants a mnemonic.

Fixes #660.  Thank you to @harlayzhou for catching and reporting!