Switch to GDI+ for BMP files
Fixes ("well enough") #663.  Thank you to @IPeluchito for catching and reporting.

FreeImage doesn't work on some BMP files.  Specifically, it reports success, but loads a blank surface at the size of the BMP file.  I haven't nailed down the specific combination of file traits that lead to this, but a quick temporary fix is to simply use GDI+ instead of FreeImage for BMP files.

Why wasn't this done in the first place?  GDI+ works well on Win 7+, but on XP, its BMP coverage is significantly worse.  These kinds of per-OS differences deterred me from using GDI+ in the early days of PhotoDemon.  (FreeImage didn't always work, but at least it didn't always work in the same, predictable ways.)

The "correct" solution here is also the nuclear option: writing my own BMP decoder.  But BMP coverage is dense (note Firefox's attempt to cover it correctly, for example: https://blog.mozilla.org/nnethercote/2015/11/06/i-rewrote-firefoxs-bmp-decoder/) and the 8 major versions/variants of the format give me pause.

I'll probably tackle it someday, but today is not that day!