JPEG-2000: implement subsampling
For the first time, this adds compatibility for a huge set of JPEG-2000 images which PD was unable to load previously.

This commit also disallows the 3rd-party FreeImage library (PD's old solution for JP2 files) from attempting to load a JP2 image if PD has rejected it.  This solves a number of potential JP2-caused crashes caused by malformed images.