New `Image > Show in file manager` menu
This command has always been available via the right-click popup menu in the image tabstrip, but if only one image is open (or the user has hidden the tabstrip) there was no way to access it.  There was also no way to assign the command a custom hotkey.

This has now been fixed!