Expand "QuickLoad" image format support to reflect recent format work
PD uses a separate "QuickLoad" function when the user adds a new layer from file (`Layer > Add > From file...`).  This is because loading images is hugely complex, and a lot of additional work can be skipped if all we need from an image is just its pixel data.

This function had not been updated against some of PD's recent format coverage work (like JPEG-2000 enhancements this past month) so this commit brings everything up-to-date for various 2025 image format enhancements.